requirements.txt
#asgiref>=3.3.1
#certifi==2020.12.5
#chardet==4.0.0
#Django==3.2.9
#django-crispy-forms==1.11.1
#django-recaptcha==2.0.6
#idna==2.10
#pytz==2021.1
#requests==2.25.1
#sqlparse==0.4.1
#urllib3==1.26.3
django-cors-headers
django-extensions
#cryptography~=3.4.6
#tqdm~=4.58.0
django-environ>=0.8.1
Pillow
#tensorflow
numpy==1.22.0
pandas
django-import-export==2.7.0
